Fix some GCC warnings regarding the sign conversion.

+++ lib/tokenize.cpp
@@ -956,7 +956,7 @@ void Tokenizer::simplifyTypedef()
         Token *argFuncRetEnd = 0;
         Token *funcStart = 0;
         Token *funcEnd = 0;
-        unsigned int offset = 1;
+        unsigned short offset = 1;
         bool function = false;
         bool functionPtr = false;
         bool functionRef = false;
